requirements.txt
yacs>=0.1.6
pyyaml>=5.1
av
matplotlib
termcolor>=1.1
simplejson
tqdm
psutil
matplotlib
detectron2@git+https://github.com/facebookresearch/detectron2.git
opencv-python
pandas
torchvision>=0.4.2
pillow
scikit-learn
tensorboard
fairscale
pytorchvideo@git+https://github.com/facebookresearch/pytorchvideo.git
pyarrow
wandb